added all

+++ 0-setup_web_static.sh
@@ -1,12 +1,42 @@
 #!/usr/bin/env bash
-# sets up the web servers for the deployment of web_static
+# sets up the webs servers for the deployment of web_static
 
-sudo apt-get -y update
-sudo apt-get -y upgrade
-sudo apt-get -y install nginx
-sudo mkdir -p /data/web_static/releases/test /data/web_static/shared
-echo "This is a test" | sudo tee /data/web_static/releases/test/index.html
-sudo ln -sf /data/web_static/releases/test/ /data/web_static/current
-sudo chown -hR ubuntu:ubuntu /data/
-sudo sed -i '38i\\tlocation /hbnb_static/ {\n\t\talias /data/web_static/current/;\n\t}\n' /etc/nginx/sites-available/default
-sudo service nginx start
+# installs nginx if not already installed
+if ! dpkg -s nginx > /dev/null
+then
+    sudo apt-get -y update
+    sudo apt-get -y install nginx
+fi
+
+mkdir -p /data/web_static/shared
+mkdir -p /data/web_static/releases/test
+
+echo '<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Test</title>
+    <style>
+        body {
+            margin: 0;
+            padding: 0;
+            text-align: center;
+        }
+    </style>
+</head>
+
+<body>
+    <h1>TEST PAGE</h1>
+</body>
+
+</html>
+
+' > /data/web_static/releases/test/index.html
+
+ln -sf /data/web_static/releases/test /data/web_static/current
+chown -R ubuntu:ubuntu /data/
+text='\\n\n\tlocation /hbnb_static {\n\t\talias /data/web_static/current;\n\t}\n'
+sed -i "24i $text" /etc/nginx/sites-enabled/default
+sudo service nginx restart
